% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized.R
\name{normalize_add_rowmeans}
\alias{normalize_add_rowmeans}
\title{Normalized Function}
\usage{
normalize_add_rowmeans(x)
}
\arguments{
\item{x}{x = raw dataset, must be a dataframe}
}
\description{
The normalized function takes the minimum and maximum values from each column and normalized
the dataset, then takes the mean of each particular wavelength, nm.
}
\keyword{absorbance}
