% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_plot.R
\name{plot_abs}
\alias{plot_abs}
\title{abs_plot Function}
\usage{
plot_abs(df, df_x, df_y, xname, yname, title)
}
\arguments{
\item{df}{df = dataset, must be in dataframe}

\item{df_x}{df_x = column of wavelength values; must be numeric}

\item{df_y}{df_y = column of desired absorbance; must be numeric}

\item{xname}{xname allows users to name the x-axis; can be character and/or numeric}

\item{yname}{yname allows users to name the y-axis; can be character and/or numeric}

\item{title}{title allows users to name the plot; can be character and/or numeric}
}
\value{
plot
}
\description{
The plot_abs function displays the absorbance of the plastic material vs wavelength from 600 to 4000 nm
including the intensity of peaks for each particular compound and molecule.
}
\keyword{absorbance}
